+++ src/Spy/Factory/SpyFactory.php
@@ -15,6 +15,8 @@ use Eloquent\Phony\Call\Factory\CallFactory;
 use Eloquent\Phony\Call\Factory\CallFactoryInterface;
 use Eloquent\Phony\Collection\IndexNormalizer;
 use Eloquent\Phony\Collection\IndexNormalizerInterface;
+use Eloquent\Phony\Invocation\Invoker;
+use Eloquent\Phony\Invocation\InvokerInterface;
 use Eloquent\Phony\Sequencer\Sequencer;
 use Eloquent\Phony\Sequencer\SequencerInterface;
 use Eloquent\Phony\Spy\Spy;
@@ -45,6 +47,7 @@ class SpyFactory implements SpyFactoryInterface
      * @param SequencerInterface|null             $labelSequencer        The label sequencer to use.
      * @param IndexNormalizerInterface|null       $indexNormalizer       The index normalizer to use.
      * @param CallFactoryInterface|null           $callFactory           The call factory to use.
+     * @param InvokerInterface|null               $invoker               The invoker to use.
      * @param TraversableSpyFactoryInterface|null $generatorSpyFactory   The generator spy factory to use.
      * @param TraversableSpyFactoryInterface|null $traversableSpyFactory The traversable spy factory to use.
      */
@@ -52,6 +55,7 @@ class SpyFactory implements SpyFactoryInterface
         SequencerInterface $labelSequencer = null,
         IndexNormalizerInterface $indexNormalizer = null,
         CallFactoryInterface $callFactory = null,
+        InvokerInterface $invoker = null,
         TraversableSpyFactoryInterface $generatorSpyFactory = null,
         TraversableSpyFactoryInterface $traversableSpyFactory = null
     ) {
@@ -64,6 +68,9 @@ class SpyFactory implements SpyFactoryInterface
         if (null === $callFactory) {
             $callFactory = CallFactory::instance();
         }
+        if (null === $invoker) {
+            $invoker = Invoker::instance();
+        }
         if (null === $generatorSpyFactory) {
             $generatorSpyFactory = GeneratorSpyFactory::instance();
         }
@@ -74,6 +81,7 @@ class SpyFactory implements SpyFactoryInterface
         $this->labelSequencer = $labelSequencer;
         $this->indexNormalizer = $indexNormalizer;
         $this->callFactory = $callFactory;
+        $this->invoker = $invoker;
         $this->generatorSpyFactory = $generatorSpyFactory;
         $this->traversableSpyFactory = $traversableSpyFactory;
     }
@@ -106,6 +114,16 @@ class SpyFactory implements SpyFactoryInterface
     public function callFactory()
     {
         return $this->callFactory;
+    }
+
+    /**
+     * Get the invoker.
+     *
+     * @return InvokerInterface The invoker.
+     */
+    public function invoker()
+    {
+        return $this->invoker;
     }
 
     /**
@@ -142,6 +160,7 @@ class SpyFactory implements SpyFactoryInterface
             strval($this->labelSequencer->next()),
             $this->indexNormalizer,
             $this->callFactory,
+            $this->invoker,
             $this->generatorSpyFactory,
             $this->traversableSpyFactory
         );
@@ -151,6 +170,7 @@ class SpyFactory implements SpyFactoryInterface
     private $labelSequencer;
     private $indexNormalizer;
     private $callFactory;
+    private $invoker;
     private $generatorSpyFactory;
     private $traversableSpyFactory;
 }