+++ test/suite/Mock/Handle/Stubbing/StubbingHandleTest.php
@@ -30,6 +30,7 @@ class StubbingHandleTest extends PHPUnit_Framework_TestCase
         $this->state = (object) array(
             'stubs' => (object) array(),
             'defaultAnswerCallback' => 'Eloquent\Phony\Stub\Stub::returnsNullAnswerCallback',
+            'isRecording' => true,
             'label' => 'label',
         );
         $this->stubFactory = new StubFactory();
@@ -252,16 +253,6 @@ EOD;
         $this->subject->noInteraction();
     }
 
-    public function testReset()
-    {
-        $this->setUpWith('Eloquent\Phony\Test\TestClassA');
-        $this->subject->stub('testClassAMethodA');
-        $this->subject->stub('testClassAMethodB');
-        $this->subject->reset();
-
-        $this->assertSame(array(), get_object_vars($this->subject->stubs()));
-    }
-
     public function testMagicCall()
     {
         $this->setUpWith('Eloquent\Phony\Test\TestClassA');
@@ -408,5 +399,34 @@ EOD;
 
         $this->assertSame('x', $this->mock->methodA('a', 'b'));
         $this->assertSame('cd', $this->mock->methodA('c', 'd'));
+    }
+
+    public function testStopRecording()
+    {
+        $this->setUpWith('Eloquent\Phony\Test\TestInterfaceA');
+
+        $this->mock->testClassAMethodA('a', 'b');
+
+        $this->assertSame($this->subject, $this->subject->stopRecording());
+
+        $this->mock->testClassAMethodB('a', 'b');
+
+        $this->subject->testClassAMethodA->called();
+        $this->subject->testClassAMethodB->never()->called();
+    }
+
+    public function testStartRecording()
+    {
+        $this->setUpWith('Eloquent\Phony\Test\TestInterfaceA');
+
+        $this->mock->testClassAMethodA('a', 'b');
+
+        $this->assertSame($this->subject, $this->subject->stopRecording());
+        $this->assertSame($this->subject, $this->subject->startRecording());
+
+        $this->mock->testClassAMethodB('a', 'b');
+
+        $this->subject->testClassAMethodA->called();
+        $this->subject->testClassAMethodB->called();
     }
 }