+++ test/suite/Spy/Factory/SpyFactoryTest.php
@@ -13,6 +13,7 @@ namespace Eloquent\Phony\Spy\Factory;
 
 use Eloquent\Phony\Call\Factory\CallFactory;
 use Eloquent\Phony\Collection\IndexNormalizer;
+use Eloquent\Phony\Invocation\Invoker;
 use Eloquent\Phony\Sequencer\Sequencer;
 use Eloquent\Phony\Spy\Spy;
 use PHPUnit_Framework_TestCase;
@@ -25,12 +26,14 @@ class SpyFactoryTest extends PHPUnit_Framework_TestCase
         $this->labelSequencer = new Sequencer();
         $this->indexNormalizer = new IndexNormalizer();
         $this->callFactory = new CallFactory();
+        $this->invoker = new Invoker();
         $this->generatorSpyFactory = new GeneratorSpyFactory();
         $this->traversableSpyFactory = new TraversableSpyFactory();
         $this->subject = new SpyFactory(
             $this->labelSequencer,
             $this->indexNormalizer,
             $this->callFactory,
+            $this->invoker,
             $this->generatorSpyFactory,
             $this->traversableSpyFactory
         );
@@ -41,6 +44,7 @@ class SpyFactoryTest extends PHPUnit_Framework_TestCase
         $this->assertSame($this->labelSequencer, $this->subject->labelSequencer());
         $this->assertSame($this->indexNormalizer, $this->subject->indexNormalizer());
         $this->assertSame($this->callFactory, $this->subject->callFactory());
+        $this->assertSame($this->invoker, $this->subject->invoker());
         $this->assertSame($this->generatorSpyFactory, $this->subject->generatorSpyFactory());
         $this->assertSame($this->traversableSpyFactory, $this->subject->traversableSpyFactory());
     }
@@ -52,6 +56,7 @@ class SpyFactoryTest extends PHPUnit_Framework_TestCase
         $this->assertSame(Sequencer::sequence('spy-label'), $this->subject->labelSequencer());
         $this->assertSame(IndexNormalizer::instance(), $this->subject->indexNormalizer());
         $this->assertSame(CallFactory::instance(), $this->subject->callFactory());
+        $this->assertSame(Invoker::instance(), $this->subject->invoker());
         $this->assertSame(GeneratorSpyFactory::instance(), $this->subject->generatorSpyFactory());
         $this->assertSame(TraversableSpyFactory::instance(), $this->subject->traversableSpyFactory());
     }
@@ -64,6 +69,7 @@ class SpyFactoryTest extends PHPUnit_Framework_TestCase
             '0',
             $this->indexNormalizer,
             $this->callFactory,
+            $this->invoker,
             $this->generatorSpyFactory,
             $this->traversableSpyFactory
         );
@@ -73,6 +79,7 @@ class SpyFactoryTest extends PHPUnit_Framework_TestCase
         $this->assertSame($callback, $actual->callback());
         $this->assertSame($this->indexNormalizer, $actual->indexNormalizer());
         $this->assertSame($this->callFactory, $actual->callFactory());
+        $this->assertSame($this->invoker, $actual->invoker());
         $this->assertSame($this->generatorSpyFactory, $actual->generatorSpyFactory());
         $this->assertSame($this->traversableSpyFactory, $actual->traversableSpyFactory());
     }